requirements.txt
# Web Framework
streamlit==1.31.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0

# Task Queue
celery[redis]==5.3.6
redis==5.0.1

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Visualization
plotly==5.18.0

# AI/ML
google-generativeai==0.3.2

# PDF Generation
reportlab==4.0.9

# HTTP Clients
requests==2.31.0
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.9
pydantic==2.5.3

# Development
pytest==7.4.4
black==24.1.1
flake8==7.0.0
